requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-jwt-extended>=4.6
python-dotenv>=1.0
